requirements.txt
flask==3.0.0
pandas==2.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
xlsxwriter==3.1.2
matplotlib==3.7.2
gunicorn==21.2.0
google-cloud-storage==2.14.0
